chatbot/requirements.txt
fastmcp
gradio
langchain
langchain-openai
langchain-mcp-adapters
duckduckgo-search
beautifulsoup4
requests
PyYAML
transformers